Update AxonServerContainerUtils
Update AxonServerContainerUtils. Since we need to use a not yet public
container, reusing the AxonServerContainerUtils is paramount for testing
 the new event storage engine. While making it public, we should move
 the purgeEvents method from the AxonServerUtils.java to this file. That
  way, we collect all axon server container operations in one place.

#3239
